ingestion/requirements.txt
requests>=2.32.0
snowflake-connector-python>=3.10.0
python-dotenv>=1.0.1
dbt-core==1.8.*
dbt-snowflake==1.8.*